KT-3576 Getters for float constants in objects always return 0.0 when constants are less than 1.0  #KT-3576 Fixed

+++ compiler/backend/src/org/jetbrains/jet/codegen/ImplementationBodyCodegen.java
@@ -1593,7 +1593,7 @@ public class ImplementationBodyCodegen extends ClassBodyCodegen {
                 if (type == Type.DOUBLE_TYPE && ((Number) value).doubleValue() == 0d) {
                     return true;
                 }
-                if (type == Type.FLOAT_TYPE && ((Number) value).byteValue() == 0f) {
+                if (type == Type.FLOAT_TYPE && ((Number) value).floatValue() == 0f) {
                     return true;
                 }
             }
